+++ CHANGES.md
@@ -23,6 +23,7 @@ Cataloging:
 -   Updated new dataset demo UI page to support adding new datasets
 -   Updated dataset ui page to be able to edit fields
 -   Added editor / editor type abstraction for making state management simpler for large forms.
+-   Added "Add Catalog" page
 
 UI:
 
@@ -31,6 +32,8 @@ UI:
 -   Removed gap after data quality star rating
 -   Replaced star emoji in static page markdown with quality star icon
 -   Refactored web client to group similar things togeather
+-   Implemented basic admin pages in react ui
+-   Started implementing new add dataset flow design changes
 
 ## 0.0.55